@@ -51,11 +51,12 @@
     "promise": "8.0.1",
     "prop-types": "^15.6.1",
     "raf": "3.4.0",
-    "react": "^16.2.0",
+    "react": "^16.3.0-alpha.2",
     "react-animations": "^1.0.0",
     "react-dev-utils": "^5.0.0",
-    "react-dom": "^16.2.0",
+    "react-dom": "^16.3.0-alpha.2",
     "react-hot-loader": "^4.0.0",
+    "react-intl": "^2.4.0",
     "react-loader": "^2.4.3",
     "react-redux": "^5.0.7",
     "react-router": "^4.2.0",
